fix: Fix generic type inference
