+++ packages/pinion/src/index.ts
@@ -3,4 +3,3 @@ export * from './core'
 export * from './ops/index'
 export * from './cli/index'
 export * from './gpt/index'
-export * from './declarations'
